Changed UIManager from Singleton Instance
The UIManager Script had a public static Instance called a singleton which was used as a global reference to the UIManager. Since the game is intended to be multiplayer there will be multiple instances of the UIManager. So I linked the UIManager to the player so that any time a player is instantiated a new UIManager reference will be created in the player script.