Added a Main Menu Scene and Audio Asset folder
I created a new Scene called Main Menu, this scene will be the starting scene which lets the player choose what to play. I also added an audio folder inside of the Assets and added the menu music to the folder. I also setup a play mission button in the Main Menu Scene. Lastly I made the main camera in the main menu play the menu music on loop.